Fixed a failing test

+++ src/TheLedgerCompany/Models/Balance.cs
@@ -5,7 +5,7 @@ namespace TheLedgerCompany.Models
         public string BankName { get; set; }
         public string BorrowerName { get; set; }
         public double AmountPaid { get; set; }
-        public double PendingEmis { get; set; }
+        public int PendingEmis { get; set; }
 
         public override string ToString()
         {
